BugID:17386144:fix the problem whitescan find

+++ kernel/rhino/pwrmgmt/cpu_tickless.c
@@ -293,7 +293,7 @@ static void tickless_enter_check(uint32_t cpu_idx, uint32_t cstate_cfg,
     if ((cStateOneShotTimer[cpu_c_state] != NULL) &&
         (cStateOneShotTimer[cpu_c_state]->one_shot_seconds_max != NULL)) {
         one_shot_max_us_support =
-          1000000 * cStateOneShotTimer[cpu_c_state]->one_shot_seconds_max();
+          1000000 * (uint64_t)cStateOneShotTimer[cpu_c_state]->one_shot_seconds_max();
     }
 
     /* if request sleeptime is longer than support, cut it down
